Placed gene ontology enrichment and KEGG pathway enrichment within the clustering module, allowing appending of GO terms to gene names displayed on graph.
Also added function to export as an expression table the values used in
the analysis after transformation. This was mainly to allow modification
by the combat batch adjustment software.

added option to turn off prduction of graph, so script can more easily
and quickly be run remotely (clustering is still calculated and the
graph is still saved as pdf).