fixed pathname bugs and data-import error
the export pathway was being assigned to the same path as the data
table directory, causing results to not be saved. Added variable
for cluster class to select cwd as default, and allow importing
of a user-defined pathway to set for all export functions.

Also updated the module functions to use the new attribute
for exporting data.