improved path filtering for GSEA
pathways were selected based on a threshold for the number of genes
in that pathway (accordingn to GO lists, etc). However, because
the data_matrix may have filtered out some of these genes due to
low expression or other reasons, the pathway now has less than
this expected number. So the pathways, once chosen, must be
checked to ensure the appropriate number of genes still remain in
the pathway for analysis.