Stabilised import of transposed tables
Some functions are still unstable with a transposed function, but
by making sure that all new cluster instances have the table in
the same orientation, then all programs downstream will function
properly. (In fact, this would probably let me clean up the
safety code that checks for the transposed matrix before working,
but it will not hurt to keep it in.

The heading line of the t-test output file has also been more
appropriately aligned.