added ignore function to q value calculation
because the hypergeometric test for gene enrichment provides a
p-value distribution with an extreme skew of 1s, the estimation
of lambda is thrown off. Two workarounds have been added:

First, it is possible to analyse q values by ignoring the p-values
which have a value of 1. Typically this leads to very small
datasets! Second, it is now possible to override pi0_hat and give
it a value of 1, which is simply it's most conservative value for
estimation of q values (and is generally more appropriate than
simply ignoring all the p-values == 1).