allowed forwarding of verbalise function
verbalise function in hicluster sits inside a function that passes
the True or False argument to determine how verbalise behaves.
The wrapper  function now returns the verbalise function, allowing
other programs to use the same functionality (or even to pass it
back to hicluster when its functions are being imported by another
script.

Also replaced the print functions within the cluster Class with
the verbalise script.