improved parsing of combinations and multithreading
combinations to be searched are now calculated before iterating
through them, so the need for checking each iteration is no longer
necessary. This has also allowed the multithreading to be fully
stacked even when the end of one file's permutations is reached,
as it just fills out iwht the next iteration on the second file.